dm-ssdcache: Make cte_match() return boolean
Most of the false positive stem from the fact that cte_match()
returns an index number. So it's impossible to tell if the
lookup failed of if the just have a cache miss.
By making cte_match() return boolean we can easily tell
if it was a cache miss, and by checking the index number
we can ever figure out if the have a cache overrun.

Signed-off-by: Hannes Reinecke <hare@suse.de>