block: Free queue resources at blk_release_queue()
A kernel crash is observed when a mounted ext3/ext4 filesystem
is physically removed.
The problem is that blk_cleanup_queue() frees up some resources
eg by calling elevator_exit(), which are not checked for in
normal operation.
So we should rather move these calls to the destructor function
blk_release_queue() as at that point all remaining references
are gone.
However, in doing so we have to ensure that any externally
supplied queue_lock is disconnected as the driver might
free up the lock after the call of blk_cleanup_queue(),

Signed-off-by: Hannes Reinecke <hare@suse.de>