dm-ssdcache: Open code retry look in process_sio()
When using list_splice() in process_sio() we shift processing
onto a private list, so it's not possible to determine how
many sios as stil in flight. This might cause quite long
delays until the list is completely processed.

So we should be open-coding the retry loop by just pulling
the first element off the list and use a retry marker to
finish processing the other elements.

Signed-off-by: Hannes Reinecke <hare@suse.de>